% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnVILBase-generics.R
\name{AnVILBase-generics}
\alias{AnVILBase-generics}
\alias{avcopy}
\alias{avlist}
\alias{avremove}
\alias{avbackup}
\alias{avrestore}
\alias{avstorage}
\title{All the generics for AnVIL}
\usage{
avcopy(..., platform)

avlist(..., platform)

avremove(..., platform)

avbackup(..., platform)

avrestore(..., platform)

avstorage(..., platform)
}
\arguments{
\item{...}{Arguments passed to the methods.}

\item{platform}{A Platform derived class indicating the AnVIL environment,
currently, \code{azure} and \code{gcp} classes are compatible.}
}
\value{
called for the side effect of registering generic functions
}
\description{
These are the generics to be used by the AnVIL package.
}
\examples{
getGeneric("avcopy")

}
\seealso{
\link{avworkspace-generics}, \link{avtable-generics}, \link{avworkflow-generics}
}
