% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnVILGenerics.R
\name{AnVILGenerics}
\alias{AnVILGenerics}
\alias{avcopy}
\alias{avlist}
\alias{avremove}
\alias{avbackup}
\alias{avrestore}
\alias{avstorage}
\alias{avworkspaces}
\alias{avtable_import}
\title{All the generics for AnVIL}
\usage{
avcopy(..., platform)

avlist(..., platform)

avremove(..., platform)

avbackup(..., platform)

avrestore(..., platform)

avstorage(..., platform)

avworkspaces(..., platform)

avtable_import(..., platform)
}
\arguments{
\item{...}{Arguments passed to the methods.}

\item{platform}{A Platform derived object.}
}
\description{
These are the generics to be used by the AnVIL package.
}
