% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avdata-generics.R
\name{avdata-generics}
\alias{avdata-generics}
\alias{avdata}
\alias{avdata_import}
\title{All the avdata generics for AnVIL}
\usage{
avdata(..., platform)

avdata_import(..., platform)
}
\arguments{
\item{...}{Arguments passed to the methods.}

\item{platform}{A Platform derived class indicating the AnVIL environment,
currently, \code{azure} and \code{gcp} classes are compatible.}
}
\value{
called for the side effect of defining the generics
}
\description{
These are the generics to be used by the AnVIL package.
}
\examples{
getGeneric("avdata")


}
\seealso{
\link{AnVILBase-generics}, \link{avworkspace-generics}, \link{avtable-generics},
\link{avworkflow-generics}
}
