% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avdata-methods.R
\name{avdata-methods}
\alias{avdata-methods}
\alias{avdata,missing-method}
\alias{avdata,ANY-method}
\alias{avdata_import,missing-method}
\alias{avdata_import,ANY-method}
\title{The avdata methods for 'missing' and 'ANY' signatures}
\usage{
\S4method{avdata}{missing}(..., platform)

\S4method{avdata}{ANY}(..., platform)

\S4method{avdata_import}{missing}(..., platform)

\S4method{avdata_import}{ANY}(..., platform)
}
\description{
The \code{AnVILBase} package defines S4 methods for the AnVIL
package. These methods are defined for the 'missing' and 'ANY' signatures.
}
\section{Functions}{
\itemize{
\item \code{avdata(missing)}: Get reference data from workspace attributes

\item \code{avdata(ANY)}: Get reference data from workspace attributes

\item \code{avdata_import(missing)}: Upload data to the workspace attributes

\item \code{avdata_import(ANY)}: Upload data to the workspace attributes

}}
