% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avnotebooks-generics.R
\name{avnotebooks-generics}
\alias{avnotebooks-generics}
\alias{avnotebooks}
\alias{avnotebooks_localize}
\alias{avnotebooks_delocalize}
\title{All the avnotebooks generics for AnVIL}
\usage{
avnotebooks(..., platform)

avnotebooks_localize(..., platform)

avnotebooks_delocalize(..., platform)
}
\arguments{
\item{...}{Arguments passed to the methods.}

\item{platform}{A Platform derived class indicating the AnVIL environment,
currently, \code{azure} and \code{gcp} classes are compatible.}
}
\value{
called for the side effect of registering generic functions
}
\description{
These are the generics to be used by the AnVIL package.
}
\examples{
getGeneric("avnotebooks")


}
\seealso{
\link{AnVILBase-generics}, \link{avworkspace-generics}, \link{avtable-generics},
\link{avdata-generics}, \link{avworkflow-generics}
}
