% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{avstop_for_status}
\alias{avstop_for_status}
\title{Check HTTP status code and raise error when appropriate}
\usage{
avstop_for_status(response, op)
}
\arguments{
\item{response}{Response object from httr}

\item{op}{Operation that was attempted}
}
\value{
\code{response} if status code less than 400 otherwise throw an error
}
\description{
This function checks the HTTP status code and raises an error
when the status code is less than 400.
}
