% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avtable-methods.R
\name{avtable-methods}
\alias{avtable-methods}
\alias{avtable_import,missing-method}
\alias{avtable_import,ANY-method}
\alias{avtable_delete_values,missing-method}
\alias{avtable_delete_values,ANY-method}
\title{The avtable methods for 'missing' and 'ANY' signatures}
\usage{
\S4method{avtable_import}{missing}(..., platform)

\S4method{avtable_import}{ANY}(..., platform)

\S4method{avtable_delete_values}{missing}(..., platform)

\S4method{avtable_delete_values}{ANY}(..., platform)
}
\description{
The \code{AnVILBase} package defines S4 methods for the AnVIL
package. These methods are defined for the 'missing' and 'ANY' signatures.
}
\section{Functions}{
\itemize{
\item \code{avtable_import(missing)}: Upload a dataset to the DATA tab

\item \code{avtable_import(ANY)}: Upload a dataset to the DATA tab

\item \code{avtable_delete_values(missing)}: Delete a row in a table based on the key

\item \code{avtable_delete_values(ANY)}: Delete a row in a table based on the key

}}
