% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avworkspace-generics.R
\name{avworkspace-generics}
\alias{avworkspace-generics}
\alias{avworkspaces}
\alias{avworkspace_namespace}
\alias{avworkspace_name}
\alias{avworkspace}
\title{All the table type generics for AnVIL}
\usage{
avworkspaces(..., platform)

avworkspace_namespace(..., platform)

avworkspace_name(..., platform)

avworkspace(..., platform)
}
\arguments{
\item{...}{Arguments passed to the methods.}

\item{platform}{A Platform derived object.}
}
\description{
These are the generics to be used by the AnVIL package.
}
\seealso{
\link{AnVILBase-generics}, \link{avtable-generics}
}
