% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_platform.R
\name{cloud_platform}
\alias{cloud_platform}
\title{Cloud Platform Identifier}
\usage{
cloud_platform()
}
\value{
A character string with the cloud platform identifier, either
"azure" or "gcp".
}
\description{
This function returns the cloud platform identifier.
}
