% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Platform-class.R
\docType{class}
\name{Platform-class}
\alias{Platform-class}
\alias{.Platform}
\title{Virtual Class for Platform}
\value{
Used mainly for inheritance.
}
\description{
This is a virtual class for Platform objects. It is used to
define the class hierarchy for Platform objects.
}
