% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnVILBase-package.R
\docType{package}
\name{AnVILBase-package}
\alias{AnVILBase}
\alias{AnVILBase-package}
\title{The AnVILBase package}
\description{
The \code{AnVILBase} package defines S4 generics for the AnVIL
package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Bioconductor/AnVILBase}
  \item Report bugs at \url{https://github.com/Bioconductor/AnVILBase/issues}
}

}
\author{
\strong{Maintainer}: Marcel Ramos \email{marcel.ramos@roswellpark.org} (\href{https://orcid.org/0000-0002-3242-0582}{ORCID})

Authors:
\itemize{
  \item Martin Morgan (\href{https://orcid.org/0000-0002-5874-8148}{ORCID}) [contributor]
}

Other contributors:
\itemize{
  \item NIH NHGRI U24HG004059 [funder]
}

}
