% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avdata-generics.R
\name{avdata-generics}
\alias{avdata-generics}
\alias{avdata}
\alias{avdata_import}
\title{All the avdata generics for AnVIL}
\usage{
avdata(..., platform)

avdata_import(..., platform)
}
\arguments{
\item{...}{Arguments passed to the methods.}

\item{platform}{A Platform derived object.}
}
\description{
These are the generics to be used by the AnVIL package.
}
\seealso{
\link{AnVILBase-generics}, \link{avworkspace-generics}, \link{avtable-generics}
}
