% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avtable-generics.R
\name{avtable-generics}
\alias{avtable-generics}
\alias{avtable}
\alias{avtables}
\alias{avtable_import}
\alias{avtable_import_set}
\alias{avtable_delete_values}
\alias{avtable_delete}
\title{All the table type generics for AnVIL}
\usage{
avtable(..., platform)

avtables(..., platform)

avtable_import(..., platform)

avtable_import_set(..., platform)

avtable_delete(..., platform)

avtable_delete_values(..., platform)
}
\arguments{
\item{...}{Arguments passed to the methods.}

\item{platform}{A Platform derived class indicating the AnVIL environment,
currently, \code{azure} and \code{gcp} classes are compatible.}
}
\value{
called for the side effect of registering generic functions
}
\description{
These are the generics to be used by the AnVIL package.
}
\examples{
getGeneric("avtable")

}
\seealso{
\link{AnVILBase-generics}, \link{avworkspace-generics}, \link{avworkflow-generics}
}
