% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avworkspace-generics.R
\name{avworkspace-generics}
\alias{avworkspace-generics}
\alias{avworkspaces}
\alias{avworkspace_namespace}
\alias{avworkspace_name}
\alias{avworkspace}
\alias{avworkspace_clone}
\title{All the table type generics for AnVIL}
\usage{
avworkspace(..., platform)

avworkspaces(..., platform)

avworkspace_name(..., platform)

avworkspace_namespace(..., platform)

avworkspace_clone(..., platform)
}
\arguments{
\item{...}{Arguments passed to the methods.}

\item{platform}{A Platform derived class indicating the AnVIL environment,
currently, \code{azure} and \code{gcp} classes are compatible.}
}
\description{
These are the generics to be used by the AnVIL package.
}
\seealso{
\link{AnVILBase-generics}, \link{avtable-generics}, \link{avdata-generics},
\link{avworkflow-generics}
}
