% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_avworkspace-generics.R
\name{has_avworkspace-generics}
\alias{has_avworkspace-generics}
\alias{has_avworkspace}
\title{The generic for checking if the user is within an environment}
\usage{
has_avworkspace(strict, ..., platform)
}
\arguments{
\item{...}{Arguments passed to the methods.}

\item{platform}{A Platform derived class indicating the AnVIL environment,
currently, \code{azure} and \code{gcp} classes are compatible.}
}
\value{
called for the side effect testing the workspace environment
}
\description{
These are generics to be used by the AnVILGCP and AnVILAz
packages.
}
\examples{
getGeneric("has_avworkspace")

}
\seealso{
\link{AnVILBase-generics}, \link{avworkspace-generics}, \link{avtable-generics}
}
